+++ project/project-tests-v8-27fca/322641929_319029211_project/utils.h
@@ -5,24 +5,24 @@
 # include <string.h>
 # include <math.h>
 
-void print_matrix(float** matrix, int dim1, int dim2);
+void print_matrix(double** matrix, int dim1, int dim2);
 
-void print_diag_matrix(float* vector, int dim);
+void print_diag_matrix(double* vector, int dim);
 
-float* init_vec_mem(int dim);
+double* init_vec_mem(int dim);
 
-float** init_matrix_mem(int dim1, int dim2);
+double** init_matrix_mem(int dim1, int dim2);
 
-void free_matrix_mem(float** matrix);
+void free_matrix_mem(double** matrix);
 
 void get_matrix_params(int *n, int *d, FILE* file);
 
 char* read_line(FILE* file, size_t* buffer_size);
 
-int parse_line_to_array(char *line, float *row, int d);
+int parse_line_to_array(char *line, double *row, int d);
 
-int init_X(FILE* file, float** X, int n, int d);
+int init_X(FILE* file, double** X, int n, int d);
 
-float** create_X(FILE* file, int n, int d);
+double** create_X(FILE* file, int n, int d);
 
 # endif
